Add files via upload
i have added all files app.py: This file represents the main Flask application. It contains the routes and logic for handling HTTP requests and responses
index.html: This file represents the HTML template for the web page. It contains the structure and layout of the web page, including forms, tables, and JavaScript code for dynamic interaction. requirements.txt is containg the background server needs as flask and database connector and dockerfile and the database(students.sql) and the compose file and at last i have added the project in a zip file.